Implement pianos first working version
This is a first working version of the task (written in README.md").
It's far away from good but it is already a good starting point to make
further developments. One of the next steps shall be to create a "Piano"
class who is in charge to manage the individual keyboard keys.